added setup for creating master password and database+table
